Upgraded Comm header
Added header_checksum, payload_checksum, and timeout_millis field; Dropped HeaderBuilder
abstraction for CommHeader

+++ src/cc/DfsBroker/Lib/Protocol.h
@@ -79,26 +79,26 @@ namespace Hypertable {
       static CommBuf *create_rename_request(const String &src,
                                             const String &dst);
 
-      virtual const char *command_text(short command);
+      virtual const char *command_text(uint64_t command);
 
-      static const uint16_t COMMAND_OPEN     = 0;
-      static const uint16_t COMMAND_CREATE   = 1;
-      static const uint16_t COMMAND_CLOSE    = 2;
-      static const uint16_t COMMAND_READ     = 3;
-      static const uint16_t COMMAND_APPEND   = 4;
-      static const uint16_t COMMAND_SEEK     = 5;
-      static const uint16_t COMMAND_REMOVE   = 6;
-      static const uint16_t COMMAND_SHUTDOWN = 7;
-      static const uint16_t COMMAND_LENGTH   = 8;
-      static const uint16_t COMMAND_PREAD    = 9;
-      static const uint16_t COMMAND_MKDIRS   = 10;
-      static const uint16_t COMMAND_STATUS   = 11;
-      static const uint16_t COMMAND_FLUSH    = 12;
-      static const uint16_t COMMAND_RMDIR    = 13;
-      static const uint16_t COMMAND_READDIR  = 14;
-      static const uint16_t COMMAND_EXISTS   = 15;
-      static const uint16_t COMMAND_RENAME   = 16;
-      static const uint16_t COMMAND_MAX      = 17;
+      static const uint64_t COMMAND_OPEN     = 0;
+      static const uint64_t COMMAND_CREATE   = 1;
+      static const uint64_t COMMAND_CLOSE    = 2;
+      static const uint64_t COMMAND_READ     = 3;
+      static const uint64_t COMMAND_APPEND   = 4;
+      static const uint64_t COMMAND_SEEK     = 5;
+      static const uint64_t COMMAND_REMOVE   = 6;
+      static const uint64_t COMMAND_SHUTDOWN = 7;
+      static const uint64_t COMMAND_LENGTH   = 8;
+      static const uint64_t COMMAND_PREAD    = 9;
+      static const uint64_t COMMAND_MKDIRS   = 10;
+      static const uint64_t COMMAND_STATUS   = 11;
+      static const uint64_t COMMAND_FLUSH    = 12;
+      static const uint64_t COMMAND_RMDIR    = 13;
+      static const uint64_t COMMAND_READDIR  = 14;
+      static const uint64_t COMMAND_EXISTS   = 15;
+      static const uint64_t COMMAND_RENAME   = 16;
+      static const uint64_t COMMAND_MAX      = 17;
 
       static const uint16_t SHUTDOWN_FLAG_IMMEDIATE = 0x0001;
